scripts/requirements.txt
scholarly
Pillow
csscompressor
jsmin
